+++ PrimalLike/Platform/PlatformWindow.cs
@@ -53,11 +53,11 @@ namespace PrimalLike.Platform
         /// <summary>
         /// Gets the width of the window.
         /// </summary>
-        public int Width { get => clientArea.Width; }
+        public uint Width { get => (uint)clientArea.Width; }
         /// <summary>
         /// Gets the height of the window.
         /// </summary>
-        public int Height { get => clientArea.Height; }
+        public uint Height { get => (uint)clientArea.Height; }
         /// <summary>
         /// Gets the aspect ratio of the window.
         /// </summary>
